+++ packages/terra-framework-docs/src/terra-dev-site/doc/date-time-picker/example/DateTimePickerDST.jsx
@@ -70,7 +70,6 @@ const DateTimePickerExampleDST = () => {
       </p>
       <Field label="Enter Date/Time" htmlFor="dst-picker">
         <DateTimePicker
-          label="Enter Date/Time"
           name="date-time-picker-example"
           dateInputAttributes={{ id: 'dst-picker' }}
           onChange={handleDateTimeChange}