+++ packages/terra-framework-docs/src/terra-dev-site/doc/date-time-picker/example/DateTimePickerDefaultDateTime.jsx
@@ -18,7 +18,6 @@ const DateTimePickerExampleDefaultDateTime = () => {
       </p>
       <Field label="Enter Date/Time" htmlFor="defaultDateTime">
         <DateTimePicker
-          label="Enter Date/Time"
           name="date-time-picker-example"
           dateInputAttributes={{ id: 'defaultDateTime' }}
           onChange={handleDateTimeChange}