+++ packages/terra-framework-docs/src/terra-dev-site/doc/date-time-picker/example/DateTimePickerMinMax.jsx
@@ -18,7 +18,6 @@ const DateTimePickerExampleMinMax = () => {
       </p>
       <Field label="Enter Date/Time" htmlFor="min-max-dates">
         <DateTimePicker
-          label="Enter Date/Time"
           name="date-time-picker-example"
           dateInputAttributes={{ id: 'min-max-dates' }}
           onChange={handleDateTimeChange}