+++ packages/terra-framework-docs/src/terra-dev-site/doc/date-time-picker/example/DateTimePickerTimeVariant.jsx
@@ -19,7 +19,6 @@ const DateTimePickerExampleTimeVariant = () => {
       </p>
       <Field label="Enter Date/Time" htmlFor="timeVariant">
         <DateTimePicker
-          label="Enter Date/Time"
           name="date-time-picker-example"
           dateInputAttributes={{ id: 'timeVariant' }}
           onChange={handleDateTimeChange}