+++ packages/terra-framework-docs/src/terra-dev-site/doc/date-time-picker/example/DateTimePickerTimeZone.jsx
@@ -27,7 +27,6 @@ function DateTimePickerTimeZone() {
       </p>
       <Field label="Enter Date/Time" htmlFor="timeZone">
         <DateTimePicker
-          label="Enter Date/Time"
           name="date-time-picker-example"
           dateInputAttributes={{ id: 'timezone' }}
           onChange={handleDateTimeChange}