+++ packages/terra-framework-docs/src/terra-dev-site/doc/date-time-picker/example/DateTimePickerWithSecondsAndTimeVariant.jsx
@@ -19,7 +19,6 @@ const DateTimePickerExampleDateWithSecondsAndTimeVariant = () => {
       </p>
       <Field label="Enter Date/Time" htmlFor="defaultDateTimeVariant">
         <DateTimePicker
-          label="Enter Date/Time"
           name="date-time-picker-example"
           dateInputAttributes={{ id: 'defaultDateTimeVariant' }}
           onChange={handleDateTimeChange}